further refactor changes
