+++ lib/cluster/cluster_master.js
@@ -29,7 +29,7 @@ module.exports = function (context) {
     messaging.register({ event: 'process:SIGINT', callback: () => {} });
 
     // event is fired from terafoundation when an error occurs during instantiation of a client
-    events.on('foundation:getConnection:error', terminalShutdown);
+    events.on('foundation:get_connection:error', terminalShutdown);
 
     function terminalShutdown(errEV) {
         logger.error(`Terminal error: ${errEV.error}`);
