+++ lib/cluster/execution_controller/engine.js
@@ -7,20 +7,18 @@ const parseError = require('error_parser');
 const moment = require('moment');
 const _ = require('lodash');
 
-module.exports = function module(executionModules) {
-    const context = executionModules.context;
-    const messaging = executionModules.messaging;
-    const executionAnalytics = executionModules.executionAnalytics;
-    const exStore = executionModules.exStore;
-    const stateStore = executionModules.stateStore;
-    const executionConfig = executionModules.executionContext.config;
-    const slicerAnalytics = executionModules.slicerAnalytics;
-
+module.exports = function module(context, messaging, exStore, stateStore, executionContext) {
+    const executionConfig = executionContext.config;
     const start = moment();
     const events = context.apis.foundation.getSystemEvents();
     const exId = process.env.ex_id;
     const jobId = process.env.job_id;
     const logger = context.apis.foundation.makeLogger({ module: 'execution_engine', ex_id: exId, job_id: jobId });
+    const slicerModule = executionContext.slicer;
+
+    let executionAnalytics;
+    let slicerAnalytics;
+    let recovery;
 
     const workerQueue = new Queue();
     const slicerQueue = new Queue();
@@ -34,8 +32,154 @@ module.exports = function module(executionModules) {
     let totalSlicers = 0;
     let allSlicersDone = false;
     let dynamicQueueLength = false;
-
+    let workerFound = false;
+    let isShuttingDown = false;
     let queueLength;
+
+    messaging.register({
+        event: 'cluster:execution:pause',
+        callback: (msg) => {
+            pause();
+            events.emit('execution:pause');
+            messaging.respond(msg);
+        }
+    });
+
+    messaging.register({
+        event: 'cluster:execution:resume',
+        callback: (msg) => {
+            resume();
+            events.emit('execution:resume');
+            messaging.respond(msg);
+        }
+    });
+
+    messaging.register({
+        event: 'worker:ready',
+        identifier: 'worker_id',
+        callback: (msg, workerId) => {
+            const workerOnlineData = msg.payload;
+            workerFound = true;
+            logger.info(`worker: ${workerId} has joined slicer: ${exId}`);
+            executionAnalytics.increment('workers_joined');
+            // if there are more clients than the length of the queue, increase the queue size
+            // TODO move queue related fn's to be used internally to engine
+            adjustQueueLength();
+            enqueueWorker(workerOnlineData);
+        }
+    });
+
+    messaging.register({
+        event: 'worker:slice:complete',
+        identifier: 'worker_id',
+        callback: (msg, workerId) => {
+            const sliceData = msg.payload;
+            executionAnalytics.increment('processed');
+            // Need to join room if a restart happened
+            if (sliceData.retry) {
+                logger.warn(`worker: ${workerId} has rejoined slicer: ${exId}`);
+                executionAnalytics.increment('workers_reconnected');
+            }
+
+            if (sliceData.error) {
+                logger.error(`worker: ${workerId} has error on slice: ${JSON.stringify(sliceData)} , slicer: ${exId}`);
+                executionAnalytics.increment('failed');
+                events.emit('slice:failure', sliceData);
+            } else {
+                if (sliceData.analytics) {
+                    slicerAnalytics.addStats(sliceData.analytics);
+                }
+                events.emit('slice:success', sliceData);
+            }
+
+            logger.debug(`worker ${workerId} has completed its slice,`, sliceData);
+
+            // Report to worker that it has been reported, so it can remove reference to last slice
+            messaging.send({
+                to: 'worker',
+                address: workerId,
+                message: 'slicer:slice:recorded',
+                payload: null
+            });
+            if (!sliceData.isShuttingDown) {
+                enqueueWorker(sliceData);
+            }
+        }
+    });
+
+    messaging.register({
+        event: 'network:disconnect',
+        identifier: 'worker_id',
+        callback: (workerId) => {
+            executionAnalytics.increment('workers_disconnected');
+            logger.warn(`Worker: ${workerId} has disconnected`);
+            // if there are less clients than the length of the queue, decrease the queue size
+            adjustQueueLength();
+            removeWorker(workerId);
+            events.emit('network:disconnect', workerId);
+            startWorkerDisconnectWatchDog();
+        }
+    });
+
+    // events can be fired from anything that instantiates a client, such as stores and slicers
+    // needs to be setup before executionRunner
+    messaging.register({ event: 'worker:shutdown', callback: executionShutdown });
+
+    messaging.register({
+        event: 'assets:loaded',
+        callback: ipcMessage => events.emit('execution:assets_loaded', ipcMessage)
+    });
+
+    events.on('slicer:execution:update', (data) => {
+        logger.debug('slicer sending a execution update', data.update);
+        // this is updating the opConfig for elasticsearch start and/or end dates for ex,
+        // this assumes elasticsearch is first
+        exStore.update(exId, { operations: data.update });
+    });
+
+    events.once('slice:failure', () => {
+        logger.error(`slicer: ${exId} has encountered a processing_error`);
+        const executionStats = executionAnalytics.getAnalytics();
+        const errorMeta = exStore.failureMetaData(null, executionStats);
+        exStore.setStatus(exId, 'failing', errorMeta);
+    });
+
+    function watchDog(checkFn, timeout, errMsg, logMsg) {
+        setTimeout(() => {
+            // if after a a set time there are still no workers, it will shutdown
+            if (checkFn()) {
+                logger.error(logMsg);
+                const executionStats = executionAnalytics.getAnalytics();
+                const errorMetaData = exStore.failureMetaData(errMsg, executionStats);
+                exStore.setStatus(exId, 'failed', errorMetaData)
+                    .then(() => {
+                        messaging.send({
+                            to: 'cluster_master',
+                            message: 'execution:error:terminal',
+                            ex_id: exId,
+                            payload: { set_status: true }
+                        });
+                    });
+            }
+        }, timeout);
+    }
+
+    function startWorkerDisconnectWatchDog() {
+        const checkFn = () => !isShuttingDown && workerFound && messaging.getClientCounts() === 0;
+        const timeout = context.sysconfig.teraslice.worker_disconnect_timeout;
+        const errMsg = `all workers from slicer #${exId} have disconnected`;
+
+        watchDog(checkFn, timeout, errMsg, errMsg);
+    }
+
+    function startWorkerConnectionWatchDog() {
+        const checkFn = () => !workerFound;
+        const timeout = context.sysconfig.teraslice.slicer_timeout;
+        const errMsg = `No workers have connected to slicer in the allotted time: ${timeout} ms`;
+        const logMsg = `A worker has not connected to a slicer for ex: ${exId}, shutting down execution`;
+
+        watchDog(checkFn, timeout, errMsg, logMsg);
+    }
 
 
     function registerSlicers(slicersArray) {
@@ -49,6 +193,85 @@ module.exports = function module(executionModules) {
     }
 
     function initialize() {
+        Promise.resolve()
+            .then(engineSetup)
+            .then(executionRecovery)
+            .catch(terminalError)
+            .then(slicerInit)
+            .catch(slicerInitRetry)
+            .then(registerSlicers)
+            .catch(terminalError);
+    }
+
+    function executionRecovery() {
+        const recoveredSlices = [];
+        const recoverExecution = process.env.recover_execution;
+
+        return new Promise((resolve, reject) => {
+            if (recoverExecution) {
+                logger.info(`execution: ${exId} is starting in recovery mode`);
+                recovery.initialize();
+
+                Promise.resolve(recovery.recoverSlices())
+                    .then(executionStartingPoints => resolve(executionStartingPoints))
+                    .catch(err => reject(parseError(err)));
+            } else {
+                resolve(recoveredSlices);
+            }
+        });
+    }
+
+
+    function slicerInit(retryDataList) {
+        return slicerModule.newSlicer(context, executionContext, retryDataList, logger);
+    }
+
+    function slicerInitRetry(slicerError) {
+        const errMsg = parseError(slicerError);
+        let times = 1;
+        // TODO maxRetries does not exist
+        const maxRetries = executionContext.maxRetries;
+        logger.error(`Error on slicer initialization, will attempt to retry ${maxRetries} times: ${errMsg}`);
+        return new Promise(((resolve, reject) => {
+            function retry() {
+                slicerInit()
+                    .then((slicerArray) => {
+                        logger.info('slicer initialization was successful');
+                        resolve(slicerArray);
+                    })
+                    .catch((err) => {
+                        times += 1;
+                        if (times < maxRetries) {
+                            retry();
+                        } else {
+                            const retryErrMsg = parseError(err);
+                            reject(`Could not initialize slicers, error: ${retryErrMsg}`);
+                        }
+                    });
+            }
+
+            retry();
+        }));
+    }
+
+    function terminalError(err) {
+        const errMsg = parseError(err);
+        const executionStats = executionAnalytics.getAnalytics();
+        const errorMeta = exStore.failureMetaData(errMsg, executionStats);
+        logger.error(errMsg);
+
+        exStore.setStatus(exId, 'failed', errorMeta)
+            .then(() => {
+                messaging.send({
+                    to: 'cluster_master',
+                    message: 'execution:error:terminal',
+                    ex_id: exId,
+                    payload: { set_status: true }
+                });
+            });
+    }
+
+    function engineSetup() {
         engineFn = function slicerEngineExecution() {
             while (workerQueue.size() && slicerQueue.size()) {
                 const worker = workerQueue.dequeue();
@@ -80,11 +303,20 @@ module.exports = function module(executionModules) {
         logger.info(`execution: ${exId} has initialized and is listening on port ${executionConfig.slicer_port}`);
         exStore.setStatus(exId, 'running');
 
+        executionAnalytics = require('./execution_analytics')(context, messaging);
+        slicerAnalytics = require('./slice_analytics')(context, executionContext);
+        setQueueLength(executionContext);
+        recovery = require('./recovery')(context, messaging, executionAnalytics, exStore, stateStore, executionContext, enqueueSlice);
+
+        messaging.initialize({ port: executionContext.config.slicer_port });
+
         // start the engine
         if (engineCanRun) {
             logger.debug('starting the slicer engine');
             engine = setInterval(engineFn, 1);
         }
+
+        startWorkerConnectionWatchDog();
     }
 
     function _allocateSlice(sliceRequest, slicerId, slicerOrder) {
@@ -248,6 +480,30 @@ module.exports = function module(executionModules) {
                 }
             }, 100);
         });
+    }
+
+    function executionShutdown() {
+        logger.info(`slicer for execution: ${exId} has received a shutdown notice`);
+        isShuttingDown = true;
+        // TODO make shutdown consistent with how we treat other modules
+        shutdown();
+        executionAnalytics.shutdown();
+
+        events.emit('execution:stop');
+        Promise.resolve()
+            .then(() => {
+                if (stateStore) {
+                    return stateStore.shutdown();
+                }
+                return true;
+            })
+            .then(() => logger.flush())
+            .then(() => process.exit())
+            .catch((err) => {
+                const errMsg = parseError(err);
+                logger.error(errMsg);
+                process.exit();
+            });
     }
 
     function _slicerCompleted() {
