+++ lib/cluster/execution_controller/execution_analytics.js
@@ -4,9 +4,7 @@ const moment = require('moment');
 const dateFormat = require('../../utils/date_utils').dateFormat;
 const _ = require('lodash');
 
-module.exports = function module(executionModules) {
-    const context = executionModules.context;
-    const messaging = executionModules.messaging;
+module.exports = function module(context, messaging) {
     const events = context.apis.foundation.getSystemEvents();
     const exId = process.env.ex_id;
     const jobId = process.env.job_id;
