+++ lib/cluster/execution_controller/recovery.js
@@ -4,15 +4,9 @@ const Promise = require('bluebird');
 const parseError = require('error_parser');
 const _ = require('lodash');
 
-module.exports = function module(executionModules) {
-    const context = executionModules.context;
-    const messaging = executionModules.messaging;
-    const executionAnalytics = executionModules.executionAnalytics;
-    const exStore = executionModules.exStore;
-    const stateStore = executionModules.stateStore;
-    const engine = executionModules.engine;
+module.exports = function module(context, messaging, executionAnalytics, exStore, stateStore, executionContext, enqueueSlice) {
     const events = context.apis.foundation.getSystemEvents();
-    const numOfSlicersToRecover = executionModules.executionContext.config.slicers;
+    const numOfSlicersToRecover = executionContext.config.slicers;
 
     let recoverExecution = process.env.recover_execution;
     let exId = process.env.ex_id;
@@ -108,7 +102,7 @@ module.exports = function module(executionModules) {
                                 retrieveSlices(nextID);
                             }
                         } else {
-                            slices.forEach(slice => engine.enqueueSlice(slice));
+                            slices.forEach(slice => enqueueSlice(slice));
                             _waitForRecoveryBatchCompletion()
                                 .then(() => retrieveSlices(slicerID));
                         }
