+++ lib/cluster/execution_controller/slice_analytics.js
@@ -1,11 +1,10 @@
 'use strict';
 
-module.exports = function module(executionModules) {
+module.exports = function module(context, executionContext) {
     let exId = process.env.ex_id;
     let jobId = process.env.job_id;
-    const context = executionModules.context;
     const logger = context.apis.foundation.makeLogger({ module: 'slice_analytics', ex_id: exId, job_id: jobId });
-    const operations = executionModules.executionContext.config.operations;
+    const operations = executionContext.config.operations;
 
     // create a container to hold all the slice analytics for this execution
     const sliceAnalytics = { time: [], size: [], memory: [] };
