+++ lib/cluster/runners/execution.js
@@ -76,16 +76,16 @@ module.exports = function module(context) {
                 return executionApi();
             });
     }
-
+    // TODO test the event execution:assets_loaded
     function initialize(events, logger) {
         return new Promise(((resolve, reject) => {
             let gettingJob = true;
             let gettingJobInterval;
 
-            events.on('worker:assets_loaded', (ipcMessage) => {
+            events.on('execution:assets_loaded', (ipcMessage) => {
                 if (ipcMessage.error) {
                     logger.error(`Error while loading assets, error: ${ipcMessage.error}`);
-                    events.removeAllListeners('worker:assets_loaded');
+                    events.removeAllListeners('execution:assets_loaded');
                     reject(ipcMessage.error);
                 } else {
                     gettingJobInterval = setInterval(() => {
@@ -94,12 +94,12 @@ module.exports = function module(context) {
                             Promise.resolve(_instantiateJob())
                                 .then((jobInstance) => {
                                     clearInterval(gettingJobInterval);
-                                    events.removeAllListeners('worker:assets_loaded');
+                                    events.removeAllListeners('execution:assets_loaded');
                                     resolve(jobInstance);
                                 })
                                 .catch((err) => {
                                     clearInterval(gettingJobInterval);
-                                    events.removeAllListeners('worker:assets_loaded');
+                                    events.removeAllListeners('execution:assets_loaded');
                                     logger.error('error initializing execution after loading assets', err.message);
                                     reject(err.message);
                                 });
@@ -110,7 +110,7 @@ module.exports = function module(context) {
 
             Promise.resolve(_instantiateJob())
                 .then((jobInstance) => {
-                    events.removeAllListeners('worker:assets_loaded');
+                    events.removeAllListeners('execution:assets_loaded');
                     clearInterval(gettingJobInterval);
                     _registerContextAPI();
