+++ lib/cluster/runners/op.js
@@ -38,7 +38,7 @@ module.exports = function module(context) {
         } catch (err) {
             const errMsg = `No configuration for endpoint ${clientConfig.endpoint} was found in the terafoundation connectors config, error: ${err.stack}`;
             context.logger.error(errMsg);
-            events.emit('foundation:getConnection:error', { error: errMsg });
+            events.emit('foundation:get_connection:error', { error: errMsg });
             return false;
         }
     }
