+++ lib/cluster/worker.js
@@ -41,7 +41,7 @@ module.exports = function module(context) {
     const messaging = messagingFn(context, logger);
 
     // if worker cannot make client, job needs to shutdown, needs to be setup before executionRunner
-    events.on('foundation:getConnection:error', terminalShutdown);
+    events.on('foundation:get_connection:error', terminalShutdown);
 
     const host = messaging.getHostUrl();
 
@@ -49,7 +49,7 @@ module.exports = function module(context) {
 
     messaging.register({
         event: 'assets:loaded',
-        callback: ipcMessage => events.emit('worker:assets_loaded', ipcMessage)
+        callback: ipcMessage => events.emit('execution:assets_loaded', ipcMessage)
     });
 
     // set up listener
