+++ lib/utils/config.js
@@ -25,7 +25,7 @@ function getClient(context, config, type) {
     } catch (err) {
         const errMsg = `No configuration for endpoint ${clientConfig.endpoint} was found in the terafoundation connectors config, error: ${err.stack}`;
         context.logger.error(errMsg);
-        events.emit('foundation:getConnection:error', { error: errMsg });
+        events.emit('foundation:get_connection:error', { error: errMsg });
     }
 }
